requirements.txt
pandas>=2.0.0
numpy>=1.24.0
pyyaml>=6.0
hyperliquid-python-sdk>=0.19.0
scipy>=1.11.0
schedule>=1.2.0
